// src/app/tasks/enums.ts
export enum TypeEnum {
  quest = "quest",
  skill = "skill",
  diary = "diary",
  collection_log = "collection_log",
  misc = "misc"
}